make batch processing a module
Instead of a Qu::BatchJob subclass, batch processing is now a
Qu::Job::BatchProcessing module. Include the module to get batch
processing primitives.
