+++ lib/qu.rb
@@ -4,7 +4,7 @@ require 'qu/failure'
 require 'qu/hooks'
 require 'qu/payload'
 require 'qu/job'
-require 'qu/batch_job'
+require 'qu/job/batch_process'
 require 'qu/backend/base'
 require 'qu/backend/instrumented'
 require 'qu/instrumenters/noop'
